Changed ssize_t to correct int for sets in Channel class storing fd wich are int

+++ inc/Channel.hpp
@@ -6,7 +6,7 @@
 /*   By: rabril-h <rabril-h@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 19:27:20 by rabril-h          #+#    #+#             */
-/*   Updated: 2024/01/03 20:36:10 by rabril-h         ###   ########.fr       */
+/*   Updated: 2024/01/03 20:59:09 by rabril-h         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ class Channel {
 
     std::string         _channelPass; // ? The pass for thhe channel
 
-    ssize_t             _limit; // ? member limits controlled by +l MODE
+    uint64_t             _limit; // ? member limits controlled by +l MODE
     std::string         _topic; // ? topic of channel controlled by +t MODE
 
     // ? We use set here becasue
@@ -55,9 +55,9 @@ class Channel {
 
     // ? Cons We will need to iterate sets since indexing is not supported...
 
-    std::set<ssize_t>   _members; // ? number of members in the channel
-    std::set<ssize_t>   _invited; // ? number of invites in the channel set by +i MODE
-    std::set<ssize_t>   _operators; // ? number of operators in the channel set by +o MODE
+    std::set<int>   _members; // ? number of members in the channel
+    std::set<int>   _invited; // ? number of invites in the channel set by +i MODE
+    std::set<int>   _operators; // ? number of operators in the channel set by +o MODE
 
     // ? Boolean for MODES
